+++ 3rdparty/communi/src/util/irctextformat.cpp
@@ -37,12 +37,10 @@
 #include "irctextformat.h"
 #include "ircpalette.h"
 #include "irccore_p.h"
-#if QT_VERSION >= 0x050000
-#include <QRegularExpression>
-#endif
 #include <QStringList>
-#include <QRegExp>
+#include <QRegularExpression>
 #include <QUrl>
+#include <QDebug>
 #include "irc.h"
 
 IRC_BEGIN_NAMESPACE
@@ -113,15 +111,16 @@ static bool parseColors(const QString& message, int pos, int* len, int* fg = nul
         *fg = -1;
     if (bg)
         *bg = -1;
-    QRegExp rx(QLatin1String("(\\d{1,2})(?:,(\\d{1,2}))?"));
-    int idx = rx.indexIn(message, pos);
-    if (idx == pos) {
-        *len = rx.matchedLength();
+
+    QRegularExpression re("(\\d{1,2})(?:,(\\d{1,2}))?");
+    auto match = re.match(message, pos);
+    if (match.hasMatch()) {
+        *len = match.captured().length();
         if (fg)
-            *fg = rx.cap(1).toInt();
+            *fg = match.captured(1).toInt();
         if (bg) {
             bool ok = false;
-            int tmp = rx.cap(2).toInt(&ok);
+            int tmp = match.captured(2).toInt(&ok);
             if (ok)
                 *bg = tmp;
         }
@@ -139,15 +138,20 @@ static QString generateLink(const QString& protocol, const QString &raw, const Q
 static QString parseUrls(const QString& message, const QString& pattern, QList<QUrl>* urls)
 {
     QString processed = message;
-#if QT_VERSION >= 0x050000
+
     int offset = 0;
     QRegularExpression rx(pattern);
     QRegularExpressionMatchIterator it = rx.globalMatch(message);
     while (it.hasNext()) {
         QRegularExpressionMatch match = it.next();
         QString protocol;
+#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
         if (match.capturedRef(2).isEmpty()) {
             QStringRef link = match.capturedRef(1);
+#else
+        if (match.capturedView(2).isEmpty()) {
+            QStringView link = match.capturedView(1);
+#endif
             if (link.startsWith(QStringLiteral("ftp."), Qt::CaseInsensitive))
                 protocol = QStringLiteral("ftp://");
             else if (link.contains(QStringLiteral("@")))
@@ -167,32 +171,7 @@ static QString parseUrls(const QString& message, const QString& pattern, QList<Q
         if (urls)
             urls->append(QUrl(protocol + raw));
     }
-#else
-    int pos = 0;
-    QRegExp rx(pattern);
-    while ((pos = rx.indexIn(processed, pos)) >= 0) {
-        int len = rx.matchedLength();
-        QString href = processed.mid(pos, len);
-        QString raw = href;
-        raw.replace("&amp;", "&");
 
-        QString protocol;
-        if (rx.cap(2).isEmpty()) {
-            if (rx.cap(1).contains(QLatin1Char('@')))
-                protocol = QLatin1String("mailto:");
-            else if (rx.cap(1).startsWith(QLatin1String("ftp."), Qt::CaseInsensitive))
-                protocol = QLatin1String("ftp://");
-            else
-                protocol = QLatin1String("http://");
-        }
-
-        QString link = generateLink(protocol, raw, href);
-        processed.replace(pos, len, link);
-        pos += link.length();
-        if (urls)
-            urls->append(QUrl(protocol + raw));
-    }
-#endif
     return processed;
 }